+++ base/shell/explorer/precomp.h
@@ -133,7 +133,6 @@ HRESULT WINAPI _CBandSite_CreateInstance(LPUNKNOWN pUnkOuter, REFIID riid, void 
 #define TWM_OPENSTARTMENU (WM_USER + 260)
 #define TWM_SETTINGSCHANGED (WM_USER + 300)
 #define TWM_NOTIFYALLAPPBARS (WM_USER + 336)
-#define TWM_SET_Z_ORDER (WM_USER + 338)
 #define TWM_PULSE (WM_USER + 400)
 
 extern const GUID IID_IShellDesktopTray;
